A dice rolling simulator
The purpose of this program is to simulate rolling a pair of dice. When the user is ready, he/she will enter 'roll' and the program will enter a random pair of numbers between 1 and 6. If the user enters anything other than roll into the program then the program will ask the user for the correct input and restart itself.